[Misc] Split the LoRA code (#30253)
